Use a long instead of long long when converting a 64-bit pointer to a string

+++ engine/core/include/Logger.hpp
@@ -56,6 +56,6 @@ namespace birb
 	 */
 	static std::string ptr_to_str(const void* ptr)
 	{
-		return "0x" + std::to_string(reinterpret_cast<long long>(ptr));
+		return "0x" + std::to_string(reinterpret_cast<long>(ptr));
 	}
 }
